Clean up, mostly working
